# api/openapi.yaml
openapi: 3.0.3
info:
  title: PGDock API
  version: 1.0.0
  description: |
    Control-plane API for PGDock. This file is the single source of truth:
    Go server interfaces (oapi-codegen) and TypeScript client types
    (openapi-typescript) are generated from it with `make generate`.
servers:
  - url: /
paths:
...
  /api/v1/orgs:
    get:
      operationId: listOrgs
      summary: Organisations you belong to (personal first)
      tags: [orgs]
      responses:
        "200":
          description: Organisations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgList"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: createOrg
      summary: Create an organisation; you become its owner
      tags: [orgs]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrgRequest"
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Org"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/orgs/{org}:
    get:
      operationId: getOrg
      summary: One organisation
      tags: [orgs]
      parameters:
        - $ref: "#/components/parameters/OrgID"
      responses:
        "200":
          description: The organisation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Org"
        default:
          $ref: "#/components/responses/Error"
    patch:
      operationId: updateOrg
      summary: Rename an organisation or change its settings
      tags: [orgs]
      parameters:
        - $ref: "#/components/parameters/OrgID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrgRequest"
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Org"
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: deleteOrg
      summary: Delete an organisation after a 7-day grace period (owner, step-up auth)
      tags: [orgs]
      parameters:
        - $ref: "#/components/parameters/OrgID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteOrgRequest"
      responses:
        "202":
          description: The organisation is offline and will be deleted at delete_after unless an owner cancels.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgDeletion"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/orgs/{org}/members:
    get:
      operationId: listOrgMembers
      summary: Members, their roles and project memberships
      tags: [orgs]
      parameters:
        - $ref: "#/components/parameters/OrgID"
      responses:
        "200":
          description: Members.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrgMemberList"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: inviteOrgMember
      summary: Invite someone to the organisation (and optionally to projects)
      tags: [orgs]
      parameters:
        - $ref: "#/components/parameters/OrgID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteRequest"
      responses:
        "201":
          description: Invitation sent.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationCreated"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/orgs/{org}/members/{user}:
    patch:
      operationId: updateOrgMember
      summary: Change a member's organisation role
      tags: [orgs]
      parameters:
        - $ref: "#/components/parameters/OrgID"
        - $ref: "#/components/parameters/UserID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrgRoleRequest"
      responses:
        "204":
          description: Changed.
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: removeOrgMember
      summary: Remove a member; their access ends at once
      tags: [orgs]
      parameters:
        - $ref: "#/components/parameters/OrgID"
        - $ref: "#/components/parameters/UserID"
      responses:
        "204":
          description: Removed.
        default:
          $ref: "#/components/responses/Error"
  /api/v1/orgs/{org}/leave:
    post:
      operationId: leaveOrg
      summary: Leave an organisation
      tags: [orgs]
      parameters:
        - $ref: "#/components/parameters/OrgID"
      responses:
        "204":
          description: Left.
        default:
          $ref: "#/components/responses/Error"
  /api/v1/orgs/{org}/transfer-ownership:
    post:
      operationId: transferOrgOwnership
      summary: Make another member an owner and step down to admin
      tags: [orgs]
      parameters:
        - $ref: "#/components/parameters/OrgID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferOwnershipRequest"
      responses:
        "204":
          description: Transferred.
        default:
          $ref: "#/components/responses/Error"
  /api/v1/orgs/{org}/invitations:
    get:
      operationId: listOrgInvitations
      summary: Pending invitations
      tags: [orgs]
      parameters:
        - $ref: "#/components/parameters/OrgID"
      responses:
        "200":
          description: Invitations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationList"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/orgs/{org}/invitations/{invitation_id}:
    delete:
      operationId: revokeOrgInvitation
      summary: Revoke a pending invitation
      tags: [orgs]
      parameters:
        - $ref: "#/components/parameters/OrgID"
        - $ref: "#/components/parameters/InvitationID"
      responses:
        "204":
          description: Revoked.
        default:
          $ref: "#/components/responses/Error"
  /api/v1/orgs/{org}/audit:
    get:
      operationId: listOrgAudit
      summary: The organisation's audit log, newest first
      tags: [audit]
      parameters:
        - $ref: "#/components/parameters/OrgID"
        - $ref: "#/components/parameters/AuditAction"
        - $ref: "#/components/parameters/AuditOutcome"
        - $ref: "#/components/parameters/AuditTarget"
        - $ref: "#/components/parameters/AuditBefore"
        - $ref: "#/components/parameters/AuditLimit"
      responses:
        "200":
          description: Audit entries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditList"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/projects/{id}/members:
    get:
      operationId: listProjectMembers
      summary: Project members (org owners and admins are implicit admins)
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: Members.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMemberList"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: addProjectMember
      summary: Add an org member to the project, or invite someone new
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectMemberRequest"
      responses:
        "200":
          description: Added (an existing org member) or invited.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectMemberAdded"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/projects/{id}/members/{user}:
    patch:
      operationId: updateProjectMember
      summary: Change a member's project role
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/UserID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectRoleRequest"
      responses:
        "204":
          description: Changed.
        default:
          $ref: "#/components/responses/Error"
    delete:
      operationId: removeProjectMember
      summary: Remove a member from the project; their login is dropped at once
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/UserID"
      responses:
        "204":
          description: Removed.
        default:
          $ref: "#/components/responses/Error"
  /api/v1/projects/{id}/credentials:
    get:
      operationId: getMyCredentials
      summary: Your personal database login on this project, if any (no password)
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: Your login.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonalCredentialsInfo"
        default:
          $ref: "#/components/responses/Error"
    post:
      operationId: issueMyCredentials
      summary: Create or rotate your personal database login; the password is shown once
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      responses:
        "200":
          description: Your login and its new password.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PersonalCredentials"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/projects/{id}/transfer:
    post:
      operationId: transferProject
      summary: Move the project to another organisation (owner of both)
      tags: [projects]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferProjectRequest"
      responses:
        "200":
          description: Moved.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        default:
          $ref: "#/components/responses/Error"
  /api/v1/projects/{id}/audit:
    get:
      operationId: listProjectAudit
      summary: The project's slice of its organisation's audit log
      tags: [audit]
      parameters:
        - $ref: "#/components/parameters/ProjectID"
        - $ref: "#/components/parameters/AuditAction"
        - $ref: "#/components/parameters/AuditOutcome"
        - $ref: "#/components/parameters/AuditTarget"
        - $ref: "#/components/parameters/AuditBefore"
        - $ref: "#/components/parameters/AuditLimit"
      responses:
        "200":
          description: Audit entries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditList"
        default:
          $ref: "#/components/responses/Error"
...
components:
  parameters:
...
    OrgID:
      name: org
      in: path
      required: true
      schema:
        type: string
        format: uuid
    UserID:
      name: user
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    OrgList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Org"
    CreateOrgRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
    UpdateOrgRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 64
        slug:
          type: string
        members_can_create_projects:
          type: boolean
...
    OrgMemberList:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrgMember"
...
    DeleteOrgRequest:
      type: object
      required: [confirm]
      properties:
        confirm:
          type: string
          description: The organisation's name, typed.
        delete_projects:
          type: boolean
          description: Delete every project too (each gets a final backup).
    OrgDeletion:
      type: object
      required: [delete_after]
      properties:
        delete_after:
          type: string
          format: date-time
...